+++ src/core/Instructions/Operations/Operation_Subtract.cpp
@@ -6,22 +6,22 @@ namespace core::instructions::operations {
     const uint64_t value2 = data.getSources()[1].getValue(state.registers);      
     const uint64_t result = value1 - value2;
 
-    state.registers[data.getDestRegisterAddr()].set<uint32_t>(result);
+    state.registers[data.getDestRegisterAddr()].set<uint32_t>(static_cast<uint32_t>(result));
 
     state.registers.getFLAGS().setBit(
       Register::FlagIndex::Carry,
       value1 < value2
     );
 
-    const bool value1_signed = getWordBit(value1, Register::INDEX_MSB);
-    const bool value2_signed = getWordBit(value2, Register::INDEX_MSB);
-    const bool result_signed = getWordBit(result, Register::INDEX_MSB);
+    const bool value1_signed = getWordBit(static_cast<uint32_t>(value1), Register::INDEX_MSB);
+    const bool value2_signed = getWordBit(static_cast<uint32_t>(value2), Register::INDEX_MSB);
+    const bool result_signed = getWordBit(static_cast<uint32_t>(result), Register::INDEX_MSB);
     state.registers.getFLAGS().setBit(
       Register::FlagIndex::Overflow,
       value1_signed != value2_signed && value1_signed != result_signed
     );
 
-    updateFlagsZeroAndSign(result, state);
+    updateFlagsZeroAndSign(static_cast<uint32_t>(result), state);
     advancePC(data, state);
   }
 }
